fix: migrate all basic ui components
