+++ src/pages/AccountPage.tsx
@@ -7,6 +7,7 @@ import MyInfoSubSection from '@/components/sections/account/MyInfoSubSection';
 import AcademicInfoSubSection from '@/components/sections/account/AcademicInfoSubSection';
 import FavoriteDepartmentsSubSection from '@/components/sections/account/FavoriteDepartmentsSubSection';
 import { API_URL } from '@/const';
+import { Orientation } from '@/shapes/enum';
 
 const AccountPage = () => {
   const { t } = useTranslation();
@@ -16,11 +17,11 @@ const AccountPage = () => {
         <div className={classNames('section', 'section--account')}>
           <Scroller expandTop={12}>
             <MyInfoSubSection />
-            <Divider orientation={Divider.Orientation.HORIZONTAL} isVisible={true} />
+            <Divider orientation={Orientation.HORIZONTAL} isVisible={true} />
             <AcademicInfoSubSection />
-            <Divider orientation={Divider.Orientation.HORIZONTAL} isVisible={true} />
+            <Divider orientation={Orientation.HORIZONTAL} isVisible={true} />
             <FavoriteDepartmentsSubSection />
-            <Divider orientation={Divider.Orientation.HORIZONTAL} isVisible={true} />
+            <Divider orientation={Orientation.HORIZONTAL} isVisible={true} />
             <div>
               <a
                 href={`${API_URL}/session/logout?next=${window.location.origin}`}
